Calls in client now provide a protocol (http), which fixes some connection issues when running localhost.
Don't process movement input for player if there is no local player ...

Copy CombinedSMZLib to clientScripts instead of just SMZLib. Also reference combined instead of not in testpage.

Modified initial connect to return a response object that provides access control origin specification.

Moved RebuildServer to /Rebuild (there was already an index what the dilly)

Suppressed game loop in server (for now). Going to need to figure out how to get that in there with the game server and everything ...

Changed nancy port from 8080 to 1337